requirements.txt
requests>=2.31.0
langgraph>=0.0.1
openai>=2.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
